net-im/element-desktop: Fix build on ppc64
This adds a custom build step for building app.asar for ppc64, due to electron-builder not supporting architectures other than x86 and arm. This step should be self-contained, and other architectures will use electron-builder as before.

Production dependencies are currently installed again in this step, due to requiring all dependencies for typescript to compile some files in a prior step and hoisting production dependencies from the full set being a non-trivial task. A better solution might be possible, but this works for now.